feat(#102): add CustomConnectionLine
